splitted definition and implementation
