+++ src/collectors/CMakeLists.txt
@@ -2,5 +2,7 @@ target_sources(bayesmix
   PUBLIC
     base_collector.hpp
     file_collector.hpp
+    file_collector.cpp
     memory_collector.hpp
+    memory_collector.cpp
 )
