+++ src/collectors/memory_collector.cpp
@@ -1,0 +1,21 @@
+#include "memory_collector.hpp"
+
+bool MemoryCollector::next_state(google::protobuf::Message* out) {
+  curr_iter++;
+  if (curr_iter == size - 1) {
+    return false;
+  }
+  out->ParseFromString(chain[curr_iter]);
+}
+
+void MemoryCollector::collect(const google::protobuf::Message& state) {
+  std::string s;
+  state.SerializeToString(&s);
+  chain.push_back(s);
+  size++;
+}
+
+void MemoryCollector::get_state(unsigned int i,
+                                google::protobuf::Message* out) override {
+  out->ParseFromString(chain[i]);
+}
